switches to gpt-4 when encountering long messages
