+++ emergent/llms.py
@@ -3,42 +3,11 @@ import os
 import openai
 import random
 import time
-import tiktoken
 from functools import wraps
-
 from dataclasses import dataclass, field
 from typing import Optional, List
+from emergent.utils import num_tokens_from_messages
 
-def num_tokens_from_messages(messages, model="gpt-3.5-turbo-0301"):
-    """Returns the number of tokens used by a list of messages."""
-    try:
-        encoding = tiktoken.encoding_for_model(model)
-    except KeyError:
-        print("Warning: model not found. Using cl100k_base encoding.")
-        encoding = tiktoken.get_encoding("cl100k_base")
-    if model == "gpt-3.5-turbo":
-        print("Warning: gpt-3.5-turbo may change over time. Returning num tokens assuming gpt-3.5-turbo-0301.")
-        return num_tokens_from_messages(messages, model="gpt-3.5-turbo-0301")
-    elif model == "gpt-4":
-        print("Warning: gpt-4 may change over time. Returning num tokens assuming gpt-4-0314.")
-        return num_tokens_from_messages(messages, model="gpt-4-0314")
-    elif model == "gpt-3.5-turbo-0301":
-        tokens_per_message = 4  # every message follows <|start|>{role/name}\n{content}<|end|>\n
-        tokens_per_name = -1  # if there's a name, the role is omitted
-    elif model == "gpt-4-0314":
-        tokens_per_message = 3
-        tokens_per_name = 1
-    else:
-        raise NotImplementedError(f"""num_tokens_from_messages() is not implemented for model {model}. See https://github.com/openai/openai-python/blob/main/chatml.md for information on how messages are converted to tokens.""")
-    num_tokens = 0
-    for message in messages:
-        num_tokens += tokens_per_message
-        for key, value in message.items():
-            num_tokens += len(encoding.encode(value))
-            if key == "name":
-                num_tokens += tokens_per_name
-    num_tokens += 3  # every reply is primed with <|start|>assistant<|message|>
-    return num_tokens
 
 def retry_with_exponential_backoff(
     func,
@@ -141,14 +110,12 @@ def chat_gpt_prompt(func):
             raise ValueError(
                 "Returned value must be a string or emergent.Prompt object"
             )
-
-        # Calculate total tokens
-        total_tokens = num_tokens_from_messages(messages, model=model)
         
-        # Check if total tokens for request exceeds token limit
+        # Handles breach of GPT-3.5 token limit
+        total_tokens = num_tokens_from_messages(messages, model=model)
         if total_tokens > 4096:
-            logging.warning("The number of tokens in the prompt exceeds the limit (4096 tokens). Skipping this prompt.")
-            return
+            logging.warning("The number of tokens in the prompt exceeds the limit of of GPT-3.5 (4096 tokens). Temporarily switching to GPT-4.")
+            model = 'gpt-4'
             
         response = openai_chat_completion(
             model=model,
